move preprocessor from core to c

+++ c/src/test/resources/syntax/null.c
@@ -1,3 +1,5 @@
+#include "../reserved.h"
+
 struct T2{
   int *fst;
   int *snd;
@@ -10,7 +12,7 @@ struct T {
 
 int main() {
   struct T* p = (struct T*) 0;
-  unsigned long addr = &(p->link);
+  unsigned long addr = (unsigned long) &(p->link);
   ASSERT(addr == sizeof(int));
   return 0;
 }
